Change return type

+++ 0x12-more_singly_linked_lists/101-print_listint_safe.c
@@ -8,7 +8,7 @@
  */
 size_t print_listint_safe(const listint_t *head)
 {
-	long int diff;
+	size_t diff;
 	int count = 0;
 
 	while (head)
